+++ service/api/like.go
@@ -30,7 +30,7 @@ func (rt *_router) likePhoto(w http.ResponseWriter, r *http.Request, ps httprout
 	user.UserFromDatabase(dbuser)
 
 	var like Like
-	auth := checkAuthorization(r.Header.Get("Authorization"), uint64(user.ID))
+	auth := checkAuthorization(r.Header.Get("Authorization"), user.ID)
 	if auth != 0 {
 		w.WriteHeader(auth)
 		w.WriteHeader(http.StatusUnauthorized)
@@ -101,7 +101,7 @@ func (rt *_router) unlikePhoto(w http.ResponseWriter, r *http.Request, ps httpro
 	}
 	user.UserFromDatabase(dbuser)
 
-	auth := checkAuthorization(r.Header.Get("Authorization"), uint64(user.ID))
+	auth := checkAuthorization(r.Header.Get("Authorization"), user.ID)
 	if auth != 0 {
 		w.WriteHeader(auth)
 		w.WriteHeader(http.StatusUnauthorized)
